+++ R/anchor.R
@@ -13,7 +13,7 @@
 #' must be the same length.  TRUE by default.
 #' @export
 #' @examples
-#' cells <- tidytable(purpose$`NNW WNW`)
+#' cells <- tidy_table(purpose$`NNW WNW`)
 #' \dontrun{
 #'   anchor(cells, 0, 1)
 #'   anchor(cells, 1, 1:2)