+++ R/offset.R
@@ -31,8 +31,8 @@
 #' # Load some pivoted data
 #' (x <- purpose$`NNW WNW`)
 #' # Make a tidy representation
-#' cells <- tidytable(x, rownames = FALSE, colnames = FALSE)
-#' cells <- cells[!is.na(cells$character), ] # Introduce 'holes' in the data
+#' cells <- tidy_table(x)
+#' cells <- cells[!is.na(cells$chr), ] # Introduce 'holes' in the data
 #' # Select an L-shape with gaps
 #' bag <- dplyr::filter(cells, row %in% 3:4, col %in% 1:2)
 #' # Offset, notice the L has been squared-off (padded)
@@ -40,7 +40,7 @@
 #' # Select a particular cell
 #' cell <- cells[which(cells$row == 3 & cells$col == 3), ]
 #' # Offset the selection downwards, stopping before the NA.
-#' offset_S(cell, cells, boundary = ~ is.na(character))
+#' offset_S(cell, cells, boundary = ~ is.na(chr))
 #' # Offset the selection right, up to and including the fifth column.
 #' offset_E(cell, cells, boundary = ~ col == 5, include = TRUE)
 #' # Offset the selection beyond the existing cells
@@ -49,7 +49,7 @@
 #' # boundary formula on every possible cell in the given direction
 #' \dontrun{offset_E(cell, cells, boundary = ~ col == 15)}
 #' cell <- cells[which(cells$row == 7 & cells$col %in% 1:2), ]
-#' offset_N(cell, cells, boundary = ~ !is.na(character), edge = TRUE)
+#' offset_N(cell, cells, boundary = ~ !is.na(chr), edge = TRUE)
 offset <- function(bag, cells, direction, n = NULL, boundary = NULL,
                    edge = FALSE, include = FALSE) {
   test_offset_args(bag, direction, n, boundary, edge, include)