+++ R/unpivotr.R
@@ -12,9 +12,9 @@
 #' to be interpreted as data, further columns containing the 'value' of the cell
 #' are, of course, necessary for there to be any point in using this package,
 #' though they are not actually required for any of the given functions.
-#' 
+#'
 #' Data frames and matrices can be converted into a format meeting these
-#' requirements by using the \code{\link{tidytable}} function.  Excel (.xlsx)
+#' requirements by using the \code{\link{tidy_table}} function.  Excel (.xlsx)
 #' files can be imported directly into the required format with the 'tidyxl'
 #' package, available at \url{https://github.com/nacnudus/tidyxl}, which has the
 #' advantage that it retains cell formatting and comments.