+++ tests/testthat/test-anchor.R
@@ -1,7 +1,7 @@
 context("anchor()")
 
 test_that("a valid cell is returned", {
-  cells <- tidytable(purpose$`NNW WNW`)
+  cells <- tidy_table(purpose$`NNW WNW`)
   expect_error(anchor(cells, 1, 0),
                "Elements of 'rows' and 'cols' must all be >= 1")
   expect_identical(anchor(cells, 1, 1), cells[cells$row == 1 & cells$col == 1, ])
@@ -14,7 +14,7 @@ test_that("a valid cell is returned", {
 })
 
 test_that("'cross' works", {
-  cells <- tidytable(purpose$`NNW WNW`)
+  cells <- tidy_table(purpose$`NNW WNW`)
   pad <- unname(as.list(anchor(cells, 1:2, 1:3)))
   expect_equal(unlist(pad[1:2]), c(rep(c(1, 2), each = 3), rep(1:3, times = 2)))
   expect_error(anchor(cells, 1:2, 1:3, cross = FALSE),