+++ tests/testthat/test-pad.R
@@ -2,8 +2,8 @@ context("pad()")
 
 test_that("pad() returns predictable cells", {
   cells <-
-    tidytable(purpose$`NNW WNW`, FALSE, FALSE) %>%
-    dplyr::filter(!is.na(character)) # Introduce 'holes' in the data
+    tidy_table(purpose$`NNW WNW`) %>%
+    dplyr::filter(!is.na(chr)) # Introduce 'holes' in the data
   # Select a region with gaps
   bag <- dplyr::filter(cells, row %in% 3:4, col %in% 1:2)
   # Pad the gaps