+++ tests/testthat/test-tibble.R
@@ -5,8 +5,8 @@ test_that("functions return tibbles", {
   # Load some pivoted data
   x <- purpose$`NNW WNW`
   # Make a tidy representation
-  cells <- tidytable(x, FALSE, FALSE)
-  cells <- cells[!is.na(cells$character), ]
+  cells <- tidy_table(x)
+  cells <- cells[!is.na(cells$chr), ]
   # Select the cells containing the values
   datacells <-
     cells %>%
@@ -15,7 +15,7 @@ test_that("functions return tibbles", {
   col_headers <-
     cells %>%
     filter(row <= 2) %>%
-    select(row, col, header = character) %>%
+    select(row, col, header = chr) %>%
     split(.$row) # Separate each row of headers
   expect_true(tibble::is_tibble(pad(datacells)))
   expect_true(tibble::is_tibble(anchor(cells, 1, 1)))